Enhance CellInterpreter API.
As was defined, using CellInterpreters (with SimpleDataTableControls) was rather cluncky; a ColumnControlFactory
returning a SimpleDataTableControl with the approprite interpreter had to be defined, and if the interpreter itself
needed to fall back to default processing, it would need to 'copy/paste' such code within its method body. This
commit allows CellInterpreters to be configured on SimpleDataTableListeners, and allows them to return true to
trigger default cell processing.